Compositon.java
What is composition ?
Composition in Java is an object-oriented programming principle where a class contains an instance of another class as a field, establishing a "Has-A" relationship instead of an "Is-A" relationship (inheritance). This allows for better modularity, reusability, and flexibility in code design.

"Has-A" is a term used in object-oriented programming (OOP) to describe a composition relationship between classes. It means that one class contains an instance of another class as a field, rather than inheriting from it.