Create Static_matrix_multiplication.java
What is static array ?
Static array which means whose size is declare before compilation and value are assign before declaration this type of array is called static array.

A static array in Java is an array whose size is fixed at the time of declaration. Unlike dynamic data structures (like ArrayList), a static array cannot change in size after it is created. The elements of the array, however, can be modified.