Create Composition.java
Composition is a technique where another class have the object of another class, like class engine object are in class car and call the engine class from the function of the startcar object , when function of the startcar is execute the before the start function of the engine is run. 
It is the most important feature of the of the object oriented programming.